+++ crates/rpc/src/server/api.rs
@@ -234,8 +234,8 @@ impl api_server::Api for RpcService {
 
     async fn get_block_by_number(
         &self,
-        request: Request<proto::blockchain::GetBlockByNumberRequest>,
-    ) -> Result<Response<proto::blockchain::GetBlockByNumberResponse>, Status> {
+        request: Request<proto::blockchain::BlockRequest>,
+    ) -> Result<Response<proto::blockchain::Block>, Status> {
         let request = request.into_inner();
 
         debug!(target: COMPONENT, ?request);