+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.15.0 (TBD)
 
-- [BREAKING] Changed `GetBlockByNumber` to accept a `GetBlockByNumberRequest` (with optional `include_proof` flag) and returns a `GetBlockByNumberResponse` containing the block and an optional block proof ([#1864](https://github.com/0xMiden/node/pull/1864)).
+- [BREAKING] Changed `GetBlockByNumber` to accept a `BlockRequest` (with optional `include_proof` flag) and returns a response containing the block and an optional block proof ([#1864](https://github.com/0xMiden/node/pull/1864)).
 
 ## v0.14.1 (2025-04-02)